AIState; AiHelper
AIState ID byte -> int
AIState refactored from enum to class
AIHelper misc

+++ src/main/java/net/sodiumstudio/befriendmobs/template/TemplateBefriendedMob.java
@@ -40,15 +40,15 @@ public class TemplateBefriendedMob /* Your mob class */ extends PathfinderMob /*
 	// It's recommended to always keep these two values as synched data
 	protected static final EntityDataAccessor<Optional<UUID>> DATA_OWNERUUID = SynchedEntityData
 			.defineId(TemplateBefriendedMob.class/* This class */, EntityDataSerializers.OPTIONAL_UUID);
-	protected static final EntityDataAccessor<Byte> DATA_AISTATE = SynchedEntityData
-			.defineId(TemplateBefriendedMob.class/* This class */, EntityDataSerializers.BYTE);
+	protected static final EntityDataAccessor<Integer> DATA_AISTATE = SynchedEntityData
+			.defineId(TemplateBefriendedMob.class/* This class */, EntityDataSerializers.INT);
 	/* More data to sync... */
 
 	@Override
 	protected void defineSynchedData() {
 		super.defineSynchedData();
 		entityData.define(DATA_OWNERUUID, Optional.empty());
-		entityData.define(DATA_AISTATE, (byte) 0);
+		entityData.define(DATA_AISTATE, 0);
 		/* More data to sync */
 	}
 
@@ -58,7 +58,7 @@ public class TemplateBefriendedMob /* Your mob class */ extends PathfinderMob /*
 	}
 
 	@Override
-	public EntityDataAccessor<Byte> getAIStateData() {
+	public EntityDataAccessor<Integer> getAIStateData() {
 		return DATA_AISTATE;
 	}
