+++ src/main/java/net/sodiumstudio/befriendmobs/template/TemplateBefriendedMobPreset.java
@@ -42,14 +42,14 @@ public class TemplateBefriendedMobPreset extends Monster implements IBefriendedM
 
 	protected static final EntityDataAccessor<Optional<UUID>> DATA_OWNERUUID = SynchedEntityData
 			.defineId(TemplateBefriendedMobPreset.class/* CHANGE TO YOUR CLASS */, EntityDataSerializers.OPTIONAL_UUID);
-	protected static final EntityDataAccessor<Byte> DATA_AISTATE = SynchedEntityData
-			.defineId(TemplateBefriendedMobPreset.class/* CHANGE TO YOUR CLASS */, EntityDataSerializers.BYTE);
+	protected static final EntityDataAccessor<Integer> DATA_AISTATE = SynchedEntityData
+			.defineId(TemplateBefriendedMobPreset.class/* CHANGE TO YOUR CLASS */, EntityDataSerializers.INT);
 
 	@Override
 	protected void defineSynchedData() {
 		super.defineSynchedData();
 		entityData.define(DATA_OWNERUUID, Optional.empty());
-		entityData.define(DATA_AISTATE, (byte) 0);
+		entityData.define(DATA_AISTATE, 0);
 	}
 	
 	@Override
@@ -58,7 +58,7 @@ public class TemplateBefriendedMobPreset extends Monster implements IBefriendedM
 	}
 
 	@Override
-	public EntityDataAccessor<Byte> getAIStateData() {
+	public EntityDataAccessor<Integer> getAIStateData() {
 		return DATA_AISTATE;
 	}
